move mysql2 instrumentation to its own file
