+++ packages/datadog-instrumentations/index.js
@@ -3,6 +3,7 @@
 require('./src/dns')
 require('./src/memcached')
 require('./src/mysql')
+require('./src/mysql2')
 require('./src/bluebird')
 require('./src/when')
 require('./src/promise')
